refactor(mail): extract email image policy into testable module
Move processHtmlImages, isInlineImageUrl, stripCssRemoteResources,
decodeHtmlEntities, isTrackingUrl, and the TRACKER_DOMAINS list out of the
heavy EmailThread.tsx component into a new app/lib/email-image-policy.ts so
the policy can be unit-tested without pulling in the whole React component
tree (which transitively imports opentelemetry and breaks vitest).

Replaces the broken EmailThread.spec.ts (which fails at import time) with
app/lib/email-image-policy.spec.ts containing the same 4 regression tests
plus the @vitest-environment happy-dom pragma so DOMParser is available.
